Moved Register to own file
