+++ controllers/register.js
@@ -1,0 +1,34 @@
+const handleRegister= (req, res, databaseCon, bcrypt) => {
+    const {name, email, password} = req.body;
+    const saltRounds= 10; 
+    const salt = bcrypt.genSaltSync(saltRounds);
+    const hash = bcrypt.hashSync(password, salt);
+    databaseCon.transaction(trx => {
+        trx.insert({
+            hash: hash, 
+            email: email
+        })
+        .into('login')
+        .returning('email')
+        .then(loginEmail => {
+            return trx('users').returning('*').insert({
+                email: loginEmail[0].email, 
+                name: name, 
+                joined: new Date()
+    
+            }).then(user =>{
+                res.json(user[0]);
+            })
+
+        })
+        .then(trx.commit)
+        .catch(trx.rollback)
+    })
+    .catch(error =>{
+        res.status(400).json('unable to register');
+    });
+}
+
+module.exports= {
+    handleRegister: handleRegister
+};
