+++ server.js
@@ -4,6 +4,8 @@ const cors= require('cors');
 const saltRounds = 1;
 const knex = require('knex');
 
+const register= require('./controllers/register.js');
+
 const databaseCon= knex({
     client: 'pg',
     connection: {
@@ -65,36 +67,7 @@ app.post('/signin', (req, res) =>{
     
 }); 
 
-app.post('/register', (req, res) =>{
-    const {name, email, password} = req.body;
-    const saltRounds= 10; 
-    const salt = bcrypt.genSaltSync(saltRounds);
-    const hash = bcrypt.hashSync(password, salt);
-    databaseCon.transaction(trx => {
-        trx.insert({
-            hash: hash, 
-            email: email
-        })
-        .into('login')
-        .returning('email')
-        .then(loginEmail => {
-            return trx('users').returning('*').insert({
-                email: loginEmail[0].email, 
-                name: name, 
-                joined: new Date()
-    
-            }).then(user =>{
-                res.json(user[0]);
-            })
-
-        })
-        .then(trx.commit)
-        .catch(trx.rollback)
-    })
-    .catch(error =>{
-        res.status(400).json('unable to register');
-    });
-}); 
+app.post('/register', (req, res) => {register.handleRegister(req, res, databaseCon, bcrypt)}); 
 
 app.get('/profile/:id', (req, res) => {
     const {id}= req.params;
